Refaktoroi adaptProjektiToSave ja siirrä kamaa omiin tiedostoihinsa.
