+++ backend/src/handler/adaptProjektiToSaveUtil/adaptAloitusKuulutusToSave.ts
@@ -1,0 +1,17 @@
+import * as API from "../../../../common/graphql/apiModel";
+import { AloitusKuulutus } from "../../database/model";
+import { adaptHankkeenKuvausToSave, adaptIlmoituksenVastaanottajatToSave } from "./common";
+
+export function adaptAloitusKuulutusToSave(
+  aloitusKuulutus: API.AloitusKuulutusInput
+): AloitusKuulutus | null | undefined {
+  if (aloitusKuulutus) {
+    const { hankkeenKuvaus, ilmoituksenVastaanottajat, ...rest } = aloitusKuulutus;
+    return {
+      ...rest,
+      hankkeenKuvaus: adaptHankkeenKuvausToSave(hankkeenKuvaus),
+      ilmoituksenVastaanottajat: adaptIlmoituksenVastaanottajatToSave(ilmoituksenVastaanottajat),
+    };
+  }
+  return aloitusKuulutus as undefined;
+}
