+++ backend/src/handler/adaptProjektiToSaveUtil/adaptSuunnitteluSopimusToSave.ts
@@ -1,0 +1,13 @@
+import { DBProjekti } from "../../database/model";
+import * as API from "../../../../common/graphql/apiModel";
+
+export function adaptSuunnitteluSopimusToSave(
+  projekti: DBProjekti,
+  suunnitteluSopimusInput?: API.SuunnitteluSopimusInput | null
+): API.SuunnitteluSopimusInput | null | undefined {
+  if (suunnitteluSopimusInput) {
+    const { logo, ...rest } = suunnitteluSopimusInput;
+    return { ...rest, logo: logo || projekti.suunnitteluSopimus?.logo };
+  }
+  return suunnitteluSopimusInput as undefined;
+}
