+++ backend/src/handler/projektiAdapter.ts
@@ -1,30 +1,16 @@
-import {
-  Aineisto,
-  AloitusKuulutus,
-  AloitusKuulutusJulkaisu,
-  DBProjekti,
-  HyvaksymisPaatosVaihe,
-  LocalizedMap,
-  NahtavillaoloVaihe,
-  SuunnitteluVaihe,
-  Vuorovaikutus,
-  VuorovaikutusTilaisuus,
-  Yhteystieto,
-} from "../database/model";
+import { DBProjekti, Yhteystieto } from "../database/model";
 import * as API from "../../../common/graphql/apiModel";
 import { NahtavillaoloVaiheTila } from "../../../common/graphql/apiModel";
 import mergeWith from "lodash/mergeWith";
 import { KayttoOikeudetManager } from "./kayttoOikeudetManager";
 import { personSearch } from "../personSearch/personSearchClient";
 import pickBy from "lodash/pickBy";
-import remove from "lodash/remove";
 import { perustiedotValidationSchema } from "../../../src/schemas/perustiedot";
 import { ValidationError } from "yup";
 import { log } from "../logger";
 import dayjs from "dayjs";
 import { kayttoOikeudetSchema } from "../../../src/schemas/kayttoOikeudet";
 import { lisaAineistoService } from "../aineisto/lisaAineistoService";
-import { IllegalArgumentError } from "../error/IllegalArgumentError";
 import { ISO_DATE_FORMAT, parseDate } from "../util/dateUtil";
 import adaptKasittelynTila from "./adaptProjektiUtil/adaptKasittelynTila";
 import {
@@ -37,10 +23,14 @@ import {
   adaptHyvaksymisPaatosVaihe,
   adaptHyvaksymisPaatosVaiheJulkaisut,
 } from "./adaptProjektiUtil/adaptHyvaksymisPaatosVaihe";
-import { adaptHankkeenKuvaus } from "./commonAdapterUtil/adaptHankkeenKuvaus";
 import { adaptAloitusKuulutus, adaptAloitusKuulutusJulkaisut } from "./adaptProjektiUtil/adaptAloitusKuulutus";
 import { adaptSuunnitteluSopimus } from "./adaptProjektiUtil/adaptSuunitteluSopimus";
-
+import { adaptVuorovaikutusToSave } from "./adaptProjektiToSaveUtil/adaptVuorovaikutusToSave";
+import { adaptAloitusKuulutusToSave } from "./adaptProjektiToSaveUtil/adaptAloitusKuulutusToSave";
+import { adaptSuunnitteluSopimusToSave } from "./adaptProjektiToSaveUtil/adaptSuunnitteluSopimusToSave";
+import { adaptSuunnitteluVaiheToSave } from "./adaptProjektiToSaveUtil/adaptSuunnitteluVaiheToSave";
+import { adaptNahtavillaoloVaiheToSave } from "./adaptProjektiToSaveUtil/adaptNahtavillaoloVaiheToSave";
+import { adaptHyvaksymisPaatosVaiheToSave } from "./adaptProjektiToSaveUtil/adaptHyvaksymisPaatosVaiheToSave";
 export enum ProjektiEventType {
   VUOROVAIKUTUS_PUBLISHED = "VUOROVAIKUTUS_PUBLISHED",
   AINEISTO_CHANGED = "AINEISTO_CHANGED",
@@ -308,421 +298,8 @@ export class ProjektiAdapter {
   }
 }
 
-function adaptSuunnitteluVaiheToSave(
-  dbProjekti: DBProjekti,
-  suunnitteluVaihe: API.SuunnitteluVaiheInput
-): SuunnitteluVaihe {
-  function validateSuunnitteluVaihePublishing() {
-    const isSuunnitteluVaiheBeingPublished = !dbProjekti.suunnitteluVaihe?.julkinen && suunnitteluVaihe.julkinen;
-    if (isSuunnitteluVaiheBeingPublished) {
-      // Publishing is allowed only if there is a published aloituskuulutusjulkaisu
-      if (!findPublishedAloitusKuulutusJulkaisu(dbProjekti.aloitusKuulutusJulkaisut)) {
-        throw new IllegalArgumentError("Suunnitteluvaihetta ei voi julkaista ennen kuin aloituskuulutus on julkaistu");
-      }
-    }
-  }
-
-  if (
-    suunnitteluVaihe &&
-    (suunnitteluVaihe.arvioSeuraavanVaiheenAlkamisesta ||
-      suunnitteluVaihe.hankkeenKuvaus ||
-      suunnitteluVaihe.suunnittelunEteneminenJaKesto)
-  ) {
-    validateSuunnitteluVaihePublishing();
-
-    const {
-      arvioSeuraavanVaiheenAlkamisesta,
-      suunnittelunEteneminenJaKesto,
-      hankkeenKuvaus,
-      julkinen,
-      palautteidenVastaanottajat,
-    } = suunnitteluVaihe;
-    return {
-      arvioSeuraavanVaiheenAlkamisesta,
-      suunnittelunEteneminenJaKesto,
-      hankkeenKuvaus: hankkeenKuvaus ? adaptHankkeenKuvaus(hankkeenKuvaus) : undefined,
-      julkinen,
-      palautteidenVastaanottajat,
-    };
-  }
-  return undefined;
-}
-
-function adaptNahtavillaoloVaiheToSave(
-  dbNahtavillaoloVaihe: NahtavillaoloVaihe,
-  nahtavillaoloVaihe: API.NahtavillaoloVaiheInput,
-  projektiAdaptationResult: ProjektiAdaptationResult,
-  nahtavillaoloVaiheJulkaisutCount: number | undefined
-): NahtavillaoloVaihe {
-  if (!nahtavillaoloVaihe) {
-    return undefined;
-  }
-  const {
-    aineistoNahtavilla: aineistoNahtavillaInput,
-    lisaAineisto: lisaAineistoInput,
-    kuulutusYhteystiedot,
-    ilmoituksenVastaanottajat,
-    hankkeenKuvaus,
-    kuulutusPaiva,
-    kuulutusVaihePaattyyPaiva,
-    muistutusoikeusPaattyyPaiva,
-    kuulutusYhteysHenkilot,
-  } = nahtavillaoloVaihe;
-
-  const aineistoNahtavilla = adaptAineistotToSave(
-    dbNahtavillaoloVaihe?.aineistoNahtavilla,
-    aineistoNahtavillaInput,
-    projektiAdaptationResult
-  );
-
-  const lisaAineisto = adaptAineistotToSave(
-    dbNahtavillaoloVaihe?.lisaAineisto,
-    lisaAineistoInput,
-    projektiAdaptationResult
-  );
-
-  let id = dbNahtavillaoloVaihe?.id;
-  if (!id) {
-    if (nahtavillaoloVaiheJulkaisutCount) {
-      id = nahtavillaoloVaiheJulkaisutCount + 1;
-    } else {
-      id = 1;
-    }
-  }
-
-  return mergeWith({}, dbNahtavillaoloVaihe, {
-    kuulutusPaiva,
-    kuulutusVaihePaattyyPaiva,
-    muistutusoikeusPaattyyPaiva,
-    kuulutusYhteysHenkilot,
-    id,
-    aineistoNahtavilla,
-    lisaAineisto,
-    kuulutusYhteystiedot: adaptYhteystiedotToSave(kuulutusYhteystiedot),
-    ilmoituksenVastaanottajat: adaptIlmoituksenVastaanottajatToSave(ilmoituksenVastaanottajat),
-    hankkeenKuvaus: adaptHankkeenKuvausToSave(hankkeenKuvaus),
-  } as NahtavillaoloVaihe);
-}
-
-function adaptHyvaksymisPaatosVaiheToSave(
-  dbHyvaksymisPaatosVaihe: HyvaksymisPaatosVaihe,
-  hyvaksymisPaatosVaihe: API.HyvaksymisPaatosVaiheInput,
-  projektiAdaptationResult: ProjektiAdaptationResult,
-  hyvaksymisPaatosVaiheJulkaisutCount: number | undefined
-): HyvaksymisPaatosVaihe {
-  if (!hyvaksymisPaatosVaihe) {
-    return undefined;
-  }
-  const {
-    hyvaksymisPaatos: hyvaksymisPaatosInput,
-    aineistoNahtavilla: aineistoNahtavillaInput,
-    kuulutusYhteystiedot,
-    ilmoituksenVastaanottajat,
-    kuulutusPaiva,
-    kuulutusVaihePaattyyPaiva,
-    kuulutusYhteysHenkilot,
-    hallintoOikeus,
-  } = hyvaksymisPaatosVaihe;
-
-  const aineistoNahtavilla = adaptAineistotToSave(
-    dbHyvaksymisPaatosVaihe?.aineistoNahtavilla,
-    aineistoNahtavillaInput,
-    projektiAdaptationResult
-  );
-
-  const hyvaksymisPaatos = adaptAineistotToSave(
-    dbHyvaksymisPaatosVaihe?.hyvaksymisPaatos,
-    hyvaksymisPaatosInput,
-    projektiAdaptationResult
-  );
-
-  let id = dbHyvaksymisPaatosVaihe?.id;
-  if (!id) {
-    if (hyvaksymisPaatosVaiheJulkaisutCount) {
-      id = hyvaksymisPaatosVaiheJulkaisutCount + 1;
-    } else {
-      id = 1;
-    }
-  }
-
-  const newChanges: HyvaksymisPaatosVaihe = {
-    kuulutusPaiva,
-    kuulutusVaihePaattyyPaiva,
-    kuulutusYhteysHenkilot,
-    id,
-    hyvaksymisPaatos,
-    aineistoNahtavilla,
-    kuulutusYhteystiedot: adaptYhteystiedotToSave(kuulutusYhteystiedot),
-    ilmoituksenVastaanottajat: adaptIlmoituksenVastaanottajatToSave(ilmoituksenVastaanottajat),
-    hallintoOikeus,
-  };
-  return mergeWith({}, dbHyvaksymisPaatosVaihe, newChanges);
-}
-
-function adaptYhteystiedotToSave(yhteystietoInputs: Array<API.YhteystietoInput>) {
-  return yhteystietoInputs?.length > 0 ? yhteystietoInputs.map((yt) => ({ ...yt })) : undefined;
-}
-
-function adaptKayttajatunnusList(
-  projekti: DBProjekti,
-  yhteysHenkilot: Array<string>,
-  doNotForceProjektipaallikko?: boolean
-): string[] | undefined {
-  if (!yhteysHenkilot || yhteysHenkilot.length == 0) {
-    return undefined;
-  }
-  // Include only usernames that can be found from projekti
-  const unfilteredList = yhteysHenkilot.map((yh) => {
-    const projektiUser = projekti.kayttoOikeudet.find((value) => value.kayttajatunnus == yh);
-    if (projektiUser) {
-      return yh;
-    } else {
-      return undefined;
-    }
-  });
-
-  // Users with PROJEKTIPAALLIKKO role should always be in the kayttajaTunnusList
-  // Push PROJEKTIPAALLIKKO into the list if not there already
-  const projektipaallikkonTunnus = projekti.kayttoOikeudet?.find(
-    ({ rooli }) => rooli === API.ProjektiRooli.PROJEKTIPAALLIKKO
-  )?.kayttajatunnus;
-  if (!doNotForceProjektipaallikko && !unfilteredList.includes(projektipaallikkonTunnus)) {
-    unfilteredList.push(projektipaallikkonTunnus);
-  }
-
-  const list = unfilteredList.filter((yh) => yh);
-  return list.length > 0 ? list : undefined;
-}
-
-function adaptVuorovaikutusTilaisuudetToSave(
-  projekti: DBProjekti,
-  vuorovaikutusTilaisuudet: Array<API.VuorovaikutusTilaisuusInput>
-): VuorovaikutusTilaisuus[] | undefined {
-  return vuorovaikutusTilaisuudet?.length > 0
-    ? vuorovaikutusTilaisuudet.map((vv) => ({
-        ...vv,
-        esitettavatYhteystiedot: adaptYhteystiedotToSave(vv.esitettavatYhteystiedot),
-        projektiYhteysHenkilot: adaptKayttajatunnusList(projekti, vv.projektiYhteysHenkilot, true),
-      }))
-    : undefined;
-}
-
-function adaptVuorovaikutusToSave(
-  projekti: DBProjekti,
-  projektiAdaptationResult: Partial<ProjektiAdaptationResult>,
-  vuorovaikutusInput?: API.VuorovaikutusInput | null
-): Vuorovaikutus[] {
-  if (vuorovaikutusInput) {
-    // Prevent saving vuorovaikutus if suunnitteluvaihe is not yet saved
-    if (!projekti.suunnitteluVaihe) {
-      throw new IllegalArgumentError("Vuorovaikutusta ei voi lisätä ennen kuin suunnitteluvaihe on tallennettu");
-    }
-
-    const dbVuorovaikutus = findVuorovaikutusByNumber(projekti, vuorovaikutusInput.vuorovaikutusNumero);
-
-    const esittelyaineistot = adaptAineistotToSave(
-      dbVuorovaikutus?.esittelyaineistot,
-      vuorovaikutusInput.esittelyaineistot,
-      projektiAdaptationResult
-    );
-    const suunnitelmaluonnokset = adaptAineistotToSave(
-      dbVuorovaikutus?.suunnitelmaluonnokset,
-      vuorovaikutusInput.suunnitelmaluonnokset,
-      projektiAdaptationResult
-    );
-
-    const vuorovaikutusToSave: Vuorovaikutus = {
-      ...vuorovaikutusInput,
-      vuorovaikutusYhteysHenkilot: adaptKayttajatunnusList(projekti, vuorovaikutusInput.vuorovaikutusYhteysHenkilot),
-      vuorovaikutusTilaisuudet: adaptVuorovaikutusTilaisuudetToSave(
-        projekti,
-        vuorovaikutusInput.vuorovaikutusTilaisuudet
-      ),
-      //Jos vuorovaikutuksen ilmoituksella ei tarvitse olla viranomaisvastaanottajia, muokkaa adaptIlmoituksenVastaanottajatToSavea
-      ilmoituksenVastaanottajat: adaptIlmoituksenVastaanottajatToSave(vuorovaikutusInput.ilmoituksenVastaanottajat),
-      esitettavatYhteystiedot: adaptYhteystiedotToSave(vuorovaikutusInput.esitettavatYhteystiedot),
-      esittelyaineistot,
-      suunnitelmaluonnokset,
-    };
-
-    checkIfAineistoJulkinenChanged(vuorovaikutusToSave, dbVuorovaikutus, projektiAdaptationResult);
-
-    return [vuorovaikutusToSave];
-  }
-  return undefined;
-}
-
-function checkIfAineistoJulkinenChanged(
-  vuorovaikutusToSave: Vuorovaikutus,
-  dbVuorovaikutus: Vuorovaikutus,
-  projektiAdaptationResult: Partial<ProjektiAdaptationResult>
-) {
-  function vuorovaikutusPublishedForTheFirstTime() {
-    return !dbVuorovaikutus?.julkinen && vuorovaikutusToSave.julkinen;
-  }
-
-  function vuorovaikutusNotPublicAnymore() {
-    return dbVuorovaikutus && dbVuorovaikutus.julkinen && !vuorovaikutusToSave.julkinen;
-  }
-
-  function vuorovaikutusJulkaisuPaivaChanged() {
-    return (
-      dbVuorovaikutus?.vuorovaikutusJulkaisuPaiva &&
-      dbVuorovaikutus.vuorovaikutusJulkaisuPaiva !== vuorovaikutusToSave.vuorovaikutusJulkaisuPaiva
-    );
-  }
-
-  if (vuorovaikutusPublishedForTheFirstTime() || vuorovaikutusNotPublicAnymore()) {
-    projektiAdaptationResult.pushEvent({
-      eventType: ProjektiEventType.VUOROVAIKUTUS_PUBLISHED,
-      vuorovaikutusNumero: vuorovaikutusToSave.vuorovaikutusNumero,
-    } as VuorovaikutusPublishedEvent);
-  }
-
-  if (
-    vuorovaikutusPublishedForTheFirstTime() ||
-    vuorovaikutusNotPublicAnymore() ||
-    vuorovaikutusJulkaisuPaivaChanged()
-  ) {
-    projektiAdaptationResult.pushEvent({
-      eventType: ProjektiEventType.AINEISTO_CHANGED,
-    } as AineistoChangedEvent);
-  }
-}
-
-function pickAineistoFromInputByDocumenttiOid(aineistotInput: API.AineistoInput[], dokumenttiOid: string) {
-  const matchedElements = remove(aineistotInput, { dokumenttiOid });
-  if (matchedElements.length > 0) {
-    return matchedElements[0];
-  }
-  return undefined;
-}
-
-function adaptAineistotToSave(
-  dbAineistot: Aineisto[] | undefined,
-  aineistotInput: API.AineistoInput[] | undefined,
-  projektiAdaptationResult: Partial<ProjektiAdaptationResult>
-): Aineisto[] | undefined {
-  const resultAineistot = [];
-  let hasPendingChanges = undefined;
-
-  // Examine and update existing documents
-  if (dbAineistot) {
-    dbAineistot.forEach((dbAineisto) => {
-      const updateAineistoInput = pickAineistoFromInputByDocumenttiOid(aineistotInput, dbAineisto.dokumenttiOid);
-      if (updateAineistoInput) {
-        // Update existing one
-
-        if (dbAineisto.nimi !== updateAineistoInput.nimi) {
-          hasPendingChanges = true;
-        }
-        dbAineisto.nimi = updateAineistoInput.nimi;
-        dbAineisto.jarjestys = updateAineistoInput.jarjestys;
-        dbAineisto.kategoriaId = updateAineistoInput.kategoriaId;
-        resultAineistot.push(dbAineisto);
-      }
-      if (!updateAineistoInput && aineistotInput) {
-        dbAineisto.tila = API.AineistoTila.ODOTTAA_POISTOA;
-        resultAineistot.push(dbAineisto);
-        hasPendingChanges = true;
-      }
-    });
-  }
-
-  // Add new ones and optionally trigger import later
-  if (aineistotInput) {
-    for (const aineistoInput of aineistotInput) {
-      resultAineistot.push({
-        dokumenttiOid: aineistoInput.dokumenttiOid,
-        nimi: aineistoInput.nimi,
-        kategoriaId: aineistoInput.kategoriaId,
-        jarjestys: aineistoInput.jarjestys,
-        tila: API.AineistoTila.ODOTTAA_TUONTIA,
-      });
-      hasPendingChanges = true;
-    }
-  }
-
-  if (hasPendingChanges) {
-    projektiAdaptationResult.pushEvent({ eventType: ProjektiEventType.AINEISTO_CHANGED } as AineistoChangedEvent);
-  }
-  return resultAineistot;
-}
-
-function adaptSuunnitteluSopimusToSave(
-  projekti: DBProjekti,
-  suunnitteluSopimusInput?: API.SuunnitteluSopimusInput | null
-): API.SuunnitteluSopimusInput | null | undefined {
-  if (suunnitteluSopimusInput) {
-    const { logo, ...rest } = suunnitteluSopimusInput;
-    return { ...rest, logo: logo || projekti.suunnitteluSopimus?.logo };
-  }
-  return suunnitteluSopimusInput as undefined;
-}
-
-function adaptHankkeenKuvausToSave(hankkeenKuvaus: API.HankkeenKuvauksetInput): LocalizedMap<string> {
-  if (!hankkeenKuvaus) {
-    return undefined;
-  }
-  return { ...hankkeenKuvaus };
-}
-
-function adaptIlmoituksenVastaanottajatToSave(
-  vastaanottajat: API.IlmoituksenVastaanottajatInput | null | undefined
-): API.IlmoituksenVastaanottajat {
-  if (!vastaanottajat) {
-    return vastaanottajat as null | undefined;
-  }
-  const kunnat: API.KuntaVastaanottaja[] =
-    vastaanottajat?.kunnat?.map((kunta) => ({ __typename: "KuntaVastaanottaja", ...kunta })) || null;
-  if (!vastaanottajat?.viranomaiset || vastaanottajat.viranomaiset.length === 0) {
-    throw new IllegalArgumentError("Viranomaisvastaanottajia pitää olla vähintään yksi.");
-  }
-  const viranomaiset: API.ViranomaisVastaanottaja[] =
-    vastaanottajat?.viranomaiset?.map((viranomainen) => ({
-      __typename: "ViranomaisVastaanottaja",
-      ...viranomainen,
-    })) || null;
-  return { __typename: "IlmoituksenVastaanottajat", kunnat, viranomaiset };
-}
-
-function adaptAloitusKuulutusToSave(aloitusKuulutus: API.AloitusKuulutusInput): AloitusKuulutus | null | undefined {
-  if (aloitusKuulutus) {
-    const { hankkeenKuvaus, ilmoituksenVastaanottajat, ...rest } = aloitusKuulutus;
-    return {
-      ...rest,
-      hankkeenKuvaus: adaptHankkeenKuvausToSave(hankkeenKuvaus),
-      ilmoituksenVastaanottajat: adaptIlmoituksenVastaanottajatToSave(ilmoituksenVastaanottajat),
-    };
-  }
-  return aloitusKuulutus as undefined;
-}
-
 function removeUndefinedFields(object: API.Projekti): Partial<API.Projekti> {
   return pickBy(object, (value) => value !== undefined);
-}
-
-export function findVuorovaikutusByNumber(
-  projekti: DBProjekti,
-  vuorovaikutusNumero: number
-): Vuorovaikutus | undefined {
-  return projekti.vuorovaikutukset?.filter((value) => value.vuorovaikutusNumero == vuorovaikutusNumero).pop();
-}
-
-export function findPublishedAloitusKuulutusJulkaisu(
-  aloitusKuulutusJulkaisut: AloitusKuulutusJulkaisu[]
-): AloitusKuulutusJulkaisu | undefined {
-  return (
-    findJulkaisuByStatus(aloitusKuulutusJulkaisut, API.AloitusKuulutusTila.HYVAKSYTTY) ||
-    findJulkaisuByStatus(aloitusKuulutusJulkaisut, API.AloitusKuulutusTila.MIGROITU)
-  );
-}
-
-export function findJulkaisuByStatus<T extends { tila?: API.AloitusKuulutusTila }>(
-  aloitusKuulutusJulkaisut: T[],
-  tila: API.AloitusKuulutusTila
-): T | undefined {
-  return aloitusKuulutusJulkaisut?.filter((j) => j.tila == tila).pop();
 }
 
 export function isDateInThePast(kuulutusVaihePaattyyPaiva: string | undefined): boolean {
