+++ backend/src/handler/projektiAdapterJulkinen.ts
@@ -20,7 +20,7 @@ import {
 } from "../../../common/graphql/apiModel";
 import pickBy from "lodash/pickBy";
 import dayjs, { Dayjs } from "dayjs";
-import { findJulkaisuByStatus, findPublishedAloitusKuulutusJulkaisu, isDateInThePast } from "./projektiAdapter";
+import { isDateInThePast } from "./projektiAdapter";
 import { fileService } from "../files/fileService";
 import { log } from "../logger";
 import { parseDate } from "../util/dateUtil";
@@ -32,6 +32,7 @@ import {
 } from "./commonAdapterUtil/lisaaTypename";
 import { adaptHankkeenKuvaus } from "./commonAdapterUtil/adaptHankkeenKuvaus";
 import { adaptYhteystiedot } from "./commonAdapterUtil/adaptYhteystiedot";
+import { findJulkaisuByStatus, findPublishedAloitusKuulutusJulkaisu } from "./commonAdapterUtil/util";
 
 class ProjektiAdapterJulkinen {
   private applyStatus(projekti: API.ProjektiJulkinen) {
