+++ backend/src/util/findVuorovaikutusByNumber.ts
@@ -1,0 +1,8 @@
+import { Vuorovaikutus, DBProjekti } from "../database/model";
+
+export function findVuorovaikutusByNumber(
+  projekti: DBProjekti,
+  vuorovaikutusNumero: number
+): Vuorovaikutus | undefined {
+  return projekti.vuorovaikutukset?.filter((value) => value.vuorovaikutusNumero == vuorovaikutusNumero).pop();
+}
