+++ backend/src/vuorovaikutus/vuorovaikutusService.ts
@@ -1,4 +1,4 @@
-import { findVuorovaikutusByNumber } from "../handler/projektiAdapter";
+import { findVuorovaikutusByNumber } from "../util/findVuorovaikutusByNumber";
 import { IlmoituksenVastaanottajat } from "../../../common/graphql/apiModel";
 import { asiakirjaService } from "../asiakirja/asiakirjaService";
 import { fileService } from "../files/fileService";
